Moved more files around. Refactored Box and Vector so that Point class is independent of these two. Vector now derives from Point
