+++ lib/data/CrabsData.py
@@ -4,7 +4,7 @@ from typing import List
 
 import pandas as pd
 from datetime import datetime
-from lib.FolderStructure import FolderStructure
+from lib.infra.FolderStructure import FolderStructure
 from lib.model.Box import Box
 from lib.data.PandasWrapper import PandasWrapper
 from lib.seefloor.SeeFloor import SeeFloor
