+++ lib/infra/StreamToLogger.py
