+++ lib/model/Vector.py
@@ -7,17 +7,17 @@ import numpy
 from lib.model.Point import Point
 
 
-class Vector:
+class Vector(Point):
     # def __init__(self, point):
     #     self.x = point.x
     #     self.y = point.y
 
-    def __str__(self):
-        return str(Point(self.x, self.y))
-
     def __init__(self, x, y):
         self.x = x
         self.y = y
+
+    def __str__(self):
+        return str(Point(self.x, self.y))
 
     @staticmethod
     def vectorArrayAsString(vectorArray):
@@ -66,6 +66,3 @@ class Vector:
 
         # tan(theta) = Opposite / Adjacent
         return math.degrees(math.atan(self.y / self.x))
-
-    def asPoint(self):
-        return Point(self.x, self.y)
