+++ lib/seefloor/FramePhysics.py
@@ -1,4 +1,4 @@
-from lib.Camera import Camera
+from lib.imageProcessing.Camera import Camera
 from lib.model.Vector import Vector
 from lib.model.Point import Point
