+++ tests/test_Crab.py
@@ -2,7 +2,7 @@ import os
 from os import getcwd
 from unittest import TestCase
 
-from lib.Camera import Camera
+from lib.imageProcessing.Camera import Camera
 from lib.model.Point import Point
 from lib.data.CrabsData import CrabsData
 from lib.model.Crab import Crab
